+++ src/components/Collection/Collection.tsx
@@ -39,10 +39,11 @@ const Collection: React.FC<{
     ...CollectionQueryDefaults,
     ...collectionQuery,
   });
+  const collections = collectionResult.collections;
 
-  if (!collectionResult) return <Loading />;
+  if (!collections) return <Loading />;
 
-  const { metadata, links, publications } = collectionResult;
+  const { metadata, links, publications } = collections;
   const { itemsPerPage, numberOfItems, currentPage, title, description } =
     metadata;
   const totalItems = numberOfItems;
@@ -177,11 +178,16 @@ const Collection: React.FC<{
         </Form>
       </Flex>
       <SimpleGrid columns={1} gap="grid.l">
-        {publications.map((pub, c) => {
-          return (
-            <CollectionItem publication={pub} key={`collection-item-${c}`} />
-          );
-        })}
+        {publications
+          ? publications.map((pub, c) => {
+              return (
+                <CollectionItem
+                  publication={pub}
+                  key={`collection-item-${c}`}
+                />
+              );
+            })
+          : null}
       </SimpleGrid>
       <Pagination
         pageCount={pageCount}