+++ src/components/Layout/Layout.test.tsx
@@ -1,7 +1,7 @@
-import { render, screen } from "@testing-library/react";
+import { render } from "../../__tests__/testUtils/render";
+import { screen } from "@testing-library/react";
 import React from "react";
 import Layout from "./Layout";
-jest.mock("next/router", () => require("next-router-mock"));
 
 describe("Layout component", () => {
   beforeEach(() => {